requirements.txt
pyinstaller==6.3.0
pillow==10.0.1